pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fedhenet"
version = "0.1"
description = "FedHEONN library with homomorphic encryption and MQTT communication"
requires-python = ">=3.10"
dependencies = [
  "torch>=1.12",
  "torchvision>=0.13",
  "paho-mqtt>=1.6.1",
  "tenseal>=0.3.0",
  "tqdm>=4.0.0",
  "loguru>=0.7.0",
  "hydra-core>=1.3.2",
  "codecarbon>=3.0.5",
  "wandb>=0.22.1",
]
license = { text = "MIT" }
authors = [
  { name ="Alejandro Dopico-Castro", email = "alejandro.dopico2@udc.es"},
  { name="Iván Pérez Digón"},
]
readme = "README.md"

[project.scripts]
fedhenet = "fedhenet.cli:main"

[tool.setuptools]
packages = {find = {include = ["fedhenet*"], exclude = ["configs*", "data*"]}}

[dependency-groups]
dev = [
    "ruff>=0.13.2",
]
